pulled from upstream arithmetica

+++ src/crons/deleteRemovedGuilds.js
@@ -6,6 +6,10 @@ import Log from "../util/log.js";
 // = Copyright (c) NullDev = //
 // ========================= //
 
+const userDb = new QuickDB({
+    filePath: path.resolve("./data/user_data.sqlite"),
+});
+
 const guildDb = new QuickDB({
     filePath: path.resolve("./data/guild_data.sqlite"),
 });
@@ -17,7 +21,7 @@ const guildDb = new QuickDB({
  * @param {import("../util/client.js").default} client
  * @return {Promise<Number>}
  */
-const iterator = async function(dbObj, client){
+const iterator = async function(dbObj, dbname, client){
     let counter = 0;
     for (const guild of dbObj){
         const guildId = guild.id.replace("guild-", "");
@@ -25,7 +29,8 @@ const iterator = async function(dbObj, client){
 
         if (!discordGuild){
             Log.warn(`[CRON] Removing guild ${guildId} from DB...`);
-            await guildDb.delete(guild.id);
+            if (dbname === "guild") await guildDb.delete(guild.id);
+            if (dbname === "user") await userDb.delete(guild.id);
             ++counter;
             continue;
         }
@@ -41,9 +46,12 @@ const iterator = async function(dbObj, client){
 const deleteRemovedGuilds = async(client) => {
     Log.wait("[CRON] Removing non-existant guilds from DB...");
 
-    const removed = await iterator(await guildDb.all(), client);
+    const removed1 = await iterator(await userDb.all(), "user", client);
+    const removed2 = await iterator(await guildDb.all(), "guild", client);
 
-    Log.done(`[CRON] Cleaned up DB. Removed ${removed} guilds.`);
+    if (removed1 !== removed2) Log.warn("[CRON] Db Mismatch! Removed " + removed1 + " from user DB and " + removed2 + " from guild DB.");
+
+    Log.done(`[CRON] Cleaned up DB. Removed ${removed1} guilds.`);
 };
 
 export default deleteRemovedGuilds;
